Work around vagrant/ansible issues pushing content to controller.
The previous method kept the contents in the memory of the running instance of ansible,
but was not able to push values back to macOS host running vagrant guests when using sudo
as the example vagrantfile illustrates.

This workaround saves a copy of the content to /tmp/<filename>, does a localhost pull
of that temp file using synchronize (rsync) module, and then removes the temp file.

It works reliably from the ansible controller when a single ansible instance is running.

The running context of this is mac high sierra 10.13.5, vagrant 2.1.1, ansible 2.5.4